fix: resolve remaining bfloat16/float32 dtype mismatches; add architecture.md
Dtype fixes (all caused by float32 identity/parameter tensors mixing with
bfloat16 activations):

workspace.py:
  - Use candidates.dtype as canonical dtype; cast slot_queries.expand()
    and ne_temp to act_dtype instead of relying on slot_queries.dtype
  - torch.eye for lateral-competition mask now uses slots.dtype
  - Legacy MHA path: cast q and ne_temp to act_dtype
  - Final LayerNorm: norm(slots.float()).to(slots.dtype) pattern

anterior_cingulate.py: torch.eye for off-diagonal mask uses sim_mat.dtype
reasoning.py: torch.eye for attractor inhibition uses sim.dtype
thalamus.py: StreamAdapter/Thalamus norms use .float()→.to(dtype) pattern
neural_geometry.py: FractalAttentionBlock norms use .float()→.to(dtype)
language.py, math.py, association.py, orchestrator.py,
memory_attention.py: additional linter-applied float()→dtype fixes

docs:
  - architecture.md: full reproduction-ready specification covering
    Φ-structure, IIT Five Postulates, Bowtie GWS topology, vesicle
    lifecycle pseudocode with tensor shapes, BDNF structural plasticity,
    HFW dual-timescale learning, CALM+MoD adaptive compute, and preset table

+++ neuroslm/intelligence/memory_attention.py
@@ -65,7 +65,7 @@ class MemoryCrossAttention(nn.Module):
             mem_vals = mem_keys
 
         B, T, D = x.shape
-        h_in = self.norm(x)
+        h_in = self.norm(x.float()).to(dtype=x.dtype)
 
         # Top-K memory selection per batch element (avg over time first)
         with torch.no_grad():
